requirements.txt
python-dotenv
openai
langchain
langchain_openai
neo4j
textblob
faiss-cpu
streamlit